moved rewards/penalties helpers to helper package
