+++ beacon-chain/core/balances/rewards_penalties_test.go
@@ -8,79 +8,6 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/params"
 )
 
-func TestBaseRewardQuotient(t *testing.T) {
-	if params.BeaconConfig().BaseRewardQuotient != 1<<5 {
-		t.Errorf("BaseRewardQuotient should be 32 for these tests to pass")
-	}
-
-	tests := []struct {
-		a uint64
-		b uint64
-	}{
-		{0, 0},
-		{1e6 * 1e9, 988211},   //1M ETH staked, 9.76% interest.
-		{2e6 * 1e9, 1397542},  //2M ETH staked, 6.91% interest.
-		{5e6 * 1e9, 2209708},  //5M ETH staked, 4.36% interest.
-		{10e6 * 1e9, 3125000}, // 10M ETH staked, 3.08% interest.
-		{20e6 * 1e9, 4419417}, // 20M ETH staked, 2.18% interest.
-	}
-	for _, tt := range tests {
-		b := baseRewardQuotient(tt.a)
-		if b != tt.b {
-			t.Errorf("BaseRewardQuotient(%d) = %d, want = %d",
-				tt.a, b, tt.b)
-		}
-	}
-}
-
-func TestBaseReward(t *testing.T) {
-	tests := []struct {
-		a uint64
-		b uint64
-	}{
-		{0, 0},
-		{params.BeaconConfig().MinDepositAmount, 61},
-		{30 * 1e9, 1853},
-		{params.BeaconConfig().MaxDepositAmount, 1976},
-		{40 * 1e9, 1976},
-	}
-	for _, tt := range tests {
-		state := &pb.BeaconState{
-			ValidatorBalances: []uint64{tt.a},
-		}
-		// Assume 10M Eth staked (base reward quotient: 3237888).
-		b := baseReward(state, 0, 3237888)
-		if b != tt.b {
-			t.Errorf("BaseReward(%d) = %d, want = %d",
-				tt.a, b, tt.b)
-		}
-	}
-}
-
-func TestInactivityPenalty(t *testing.T) {
-	tests := []struct {
-		a uint64
-		b uint64
-	}{
-		{1, 2929},
-		{2, 3883},
-		{5, 6744},
-		{10, 11512},
-		{50, 49659},
-	}
-	for _, tt := range tests {
-		state := &pb.BeaconState{
-			ValidatorBalances: []uint64{params.BeaconConfig().MaxDepositAmount},
-		}
-		// Assume 10 ETH staked (base reward quotient: 3237888).
-		b := inactivityPenalty(state, 0, 3237888, tt.a)
-		if b != tt.b {
-			t.Errorf("InactivityPenalty(%d) = %d, want = %d",
-				tt.a, b, tt.b)
-		}
-	}
-}
-
 func TestFFGSrcRewardsPenalties(t *testing.T) {
 	tests := []struct {
 		voted                          []uint64
